requirements.txt
galaxy.plugin.api==0.69

# steam_network backend
protobuf==4.22.3
vdf==3.4
pyobjc-framework-CoreServices==5.1.2; sys_platform == 'darwin'
websockets==11.0.2
dataclasses-json==0.5.7

rsa==4.9 #current version probably ok.
